apps/api/pyproject.toml
[tool.poetry]
name = "validahub-api"
version = "0.1.0"
description = "ValidaHub API Server"
authors = ["ValidaHub Team"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.30.1"}
pydantic = "^2.7.4"
pydantic-settings = "^2.3.4"
sqlalchemy = "^2.0.31"
alembic = "^1.13.2"
psycopg = {extras = ["binary"], version = "^3.2.3"}
redis = "^5.0.7"
celery = "^5.4.0"
boto3 = "^1.34.141"
python-multipart = "^0.0.9"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pandas = "^2.2.2"
openpyxl = "^3.1.5"
pyyaml = "^6.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
pytest-asyncio = "^0.23.7"
pytest-cov = "^5.0.0"
httpx = "^0.27.0"
ruff = "^0.5.1"
mypy = "^1.10.1"
faker = "^26.0.0"
black = "^24.4.2"
isort = "^5.13.2"
pre-commit = "^3.7.1"

[tool.poetry.scripts]
dev = "uvicorn src.main:app --reload --port 8000"
test = "pytest"
lint = "ruff check ."
format = "black ."
typecheck = "mypy src"
check-deps = "python scripts/check_deps.py"

[tool.ruff]
target-version = "py311"
line-length = 120
select = ["E", "F", "I", "N", "W", "B", "C90", "UP", "S"]
ignore = ["E501", "B008", "S101"]

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
strict = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
asyncio_mode = "auto"

[tool.black]
line-length = 120
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  | venv
  | venv_new
  | \.git
  | \.mypy_cache
  | \.pytest_cache
  | __pycache__
)/
'''

[tool.isort]
profile = "black"
line_length = 120
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"